Moves service list item TS interface back into server lib file
