--- x-pack/plugins/apm/typings/Service.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export interface IServiceListItem {
-  serviceName: string;
-  agentName: string | undefined;
-  transactionsPerMinute: number;
-  errorsPerMinute: number;
-  avgResponseTime: number;
-}
